% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{AMRI}
\alias{AMRI}
\title{Adjusted Multi-resolution Rand Index between two clusterings}
\usage{
AMRI(labels1, labels2)
}
\arguments{
\item{labels1}{a vector of labels representing partition1 results}

\item{labels2}{a vector of labels representing partition2 results}
}
\value{
A list containing \describe{
    \item{mri}{scalar, calculated Mutlti-resolution Rand Index}
    \item{amri}{MRI adjusted for chance based on permutation model.}
}
}
\description{
Given two partitions on the save n data points, AMRI is used to evaluate the similarity between two
partitions given the two (different) resolutions. It is adapted from Adjusted Rand Index to account
for the difference in resolution
}
