% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree_functions.R
\name{asign_samples_to_paths}
\alias{asign_samples_to_paths}
\title{Assign data points to the viable path that are closest to the original clustering assignment}
\usage{
asign_samples_to_paths(labelmat, paths, weights.per.layer = NULL)
}
\arguments{
\item{labelmat}{Initial multi-resolution cluster assignment, n-by-m, corresponding to m resolutions}

\item{paths}{The viable paths}

\item{weights.per.layer}{the weights used for loss calculation (if supplied, all ones by default)}
}
\value{
A vector of length n containing the path index to which the data points are assigned to
}
\description{
Assign data points to the viable path that are closest to the original clustering assignment
}
