% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{augment_path}
\alias{augment_path}
\title{Augment the path to include on additional node as alternative to each layer}
\usage{
augment_path(paths, layers, aug.name = -1)
}
\arguments{
\item{paths}{the viable tree paths, saved in a label matrix}

\item{layers}{the layer index}

\item{aug.name}{name of the augmented cluster}
}
\value{
new paths after augmentation
}
\description{
Augment the path to include on additional node as alternative to each layer
}
