% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clust_example}
\alias{clust_example}
\title{A array of simulated multi-resolution clustering, with columns giving the labels for each clustering.}
\format{
A list of
\describe{
  \item{clusterings}{A matrix of sample by clusters, labels for each clusterings}
  \item{ref.labels}{A vector of reference labels of the samples.}
}
}
\usage{
clust_example
}
\description{
A array of simulated multi-resolution clustering, with columns giving the labels for each clustering.
}
\keyword{datasets}
