% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{consensus_clustering_weighted}
\alias{consensus_clustering_weighted}
\title{weighted consensus clustering}
\usage{
consensus_clustering_weighted(labelmat, k = NULL)
}
\arguments{
\item{labelmat}{n-by-m sample by clustering label matrix}

\item{k}{number of clusters for the final consensus clustering}
}
\value{
consensus clustering result saved as a label vector (NA is labels missing for all clusterings)
}
\description{
weighted consensus clustering
}
