% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{consensus_clustering_within_resolution}
\alias{consensus_clustering_within_resolution}
\title{Perform consensus clustering among clusterings of same number of clusters}
\usage{
consensus_clustering_within_resolution(
  labelmat,
  ks = NULL,
  sample.weighted = F,
  n.cores = NULL
)
}
\arguments{
\item{labelmat}{n-by-m sample by clustering label matrix}

\item{ks}{vector of length m, number of clustering for each clustering}

\item{n.cores}{maximum number of cores used for parallel computation}
}
\value{
a list containing \describe{
\item{labelmat.ccs}{consensus clustering results saved in a label matrix}
\item{k.idx}{idx mapping the original columns to the new columns}
}
}
\description{
Perform consensus clustering among clusterings of same number of clusters
}
