% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{construct_tree_from_labelmat}
\alias{construct_tree_from_labelmat}
\title{Construct the cluster tree from multi-resolution clustering save as a label matrix}
\usage{
construct_tree_from_labelmat(labelmat)
}
\arguments{
\item{labelmat}{sample by m label matrix, each columns represents a clustering for a certain resolution}
}
\value{
tree saved as an edge list containing \describe{
    \item{start}{label of start node}
    \item{end}{label of end node}
}
}
\description{
Construct the cluster tree from multi-resolution clustering save as a label matrix
}
