% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{create_sce_from_counts}
\alias{create_sce_from_counts}
\title{Create SingleCellExperiment object from count data}
\usage{
create_sce_from_counts(counts, col.data, row.data = NULL, scale.factor = 10000)
}
\arguments{
\item{counts}{gene-by-cell count matrix}

\item{col.data}{a dataframe containing cell informations}

\item{row.data}{a dataframe containing gene informations}

\item{scale.factor}{scalar used for per cell sum of count normalization}
}
\value{
a SingleCellExperiment object
}
\description{
Create SingleCellExperiment object from count data
}
