% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{A simulated single cell RNAseq data of 500 genes-by-500 cells, generated using SymSim simulation package}
\format{
A list of seven attributes
\describe{
\item{counts}{gene-by-cell count matrix}
\item{metadata}{information of cells}
\item{params}{simulation parameters}
\item{true_counts}{true counts without noise and dropouts}
\item{gene_len}{}
\item{tsne_true_counts}{tSNE plot of true counts}
\item{tsne_UMI_counts}{tSNE plot of observed counts}
}
}
\usage{
data_example
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
