% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{decode}
\alias{decode}
\title{parse the node name as the layer and cluster}
\usage{
decode(node.name)
}
\arguments{
\item{node.name}{name of the node}
}
\value{
A list \describe{
    \item{layer}{the layer in the tree the node belongs to}
    \item{label}{the label of the cluster in the layer the node belongs to}
}
}
\description{
parse the node name as the layer and cluster
}
