% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{diff_between_similarity_mat}
\alias{diff_between_similarity_mat}
\title{Calculate the L1 distance between two similarity matrix.
Get the diff of two trees, which are represented with the similarity matrices}
\usage{
diff_between_similarity_mat(sim.mat1, sim.mat2)
}
\arguments{
\item{sim.mat1}{first similarity matrix}

\item{sim.mat2}{second similarity matrix}
}
\value{

}
\description{
Calculate the L1 distance between two similarity matrix.
Get the diff of two trees, which are represented with the similarity matrices
}
