% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{div_0}
\alias{div_0}
\title{pointwise division of matrices such that 0/0=0}
\usage{
div_0(A, B)
}
\arguments{
\item{A}{enumerator}

\item{B}{denominator}
}
\description{
pointwise division of matrices such that 0/0=0
}
