% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{expected_MRI_perm}
\alias{expected_MRI_perm}
\title{expected Multi-resolution Rand Index (MRI) under random permutation model}
\usage{
expected_MRI_perm(labels1, labels2)
}
\arguments{
\item{labels1}{first label vector used for calculating the MRI}

\item{labels2}{second label vector used for calculating the MRI}
}
\value{
scalar of expected MRI
}
\description{
expected Multi-resolution Rand Index (MRI) under random permutation model
}
