% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{exponential_sampling}
\alias{exponential_sampling}
\title{Exponential sampling}
\usage{
exponential_sampling(n.res, res.min = 0.001, res.max = 3, base = 2)
}
\arguments{
\item{n.res}{integer, number of resolutions to sample}

\item{res.min}{scalar, lower bound}

\item{res.max}{scalar, upper bound}

\item{base}{base of the exponential scale (default 2)}
}
\value{
a vector of sampled parameters
}
\description{
Sample resolutions in exponential scale (sparser for larger values)
}
