% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_simulation_functions.R
\name{generateDataSymSim}
\alias{generateDataSymSim}
\title{Generate single cell data using SymSim model}
\usage{
generateDataSymSim(
  ncells,
  ngenes,
  tree,
  params = NULL,
  seed = 42,
  plot.tsne = F,
  ...
)
}
\arguments{
\item{ncells}{scalar, number of cells}

\item{ngenes}{scalar, number of genes}

\item{tree}{a phylo object specifying the tree struction where the cell type follow}

\item{params}{a list containing the simulation parameters used in SymSim}

\item{plot.tsne}{boolean whether to generate tsne plot}

\item{...}{other parameters passed to SymSim function}
}
\value{
A list of \describe{
\item{count}{ncell-by-ngene count matrix}
\item{params}{simulation parameters}
\item{true_counts}{ncell-by-ngene count matrix without noise and dropouts}
\item{gene_len}{}
\item{tsne_true_counts}{tSNE plot of the true counts}
\item{tsne_UMI_counts}{tSNE plot of the simulated observed counts}
\item{metadata}{Information of the cells}
}
}
\description{
Generate single cell data using SymSim model
}
