% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree_functions.R
\name{get_bad_nodeset}
\alias{get_bad_nodeset}
\title{Output the set of bad node in the given tree}
\usage{
get_bad_nodeset(tree)
}
\arguments{
\item{tree}{a list containg start & and pairs representing a cluster tree}
}
\value{
A vector containing the name of the bad vertex
}
\description{
Output the set of bad node in the given tree
}
