% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{get_index}
\alias{get_index}
\title{Calculate the index between two clusterings}
\usage{
get_index(labels1, labels2, index.name = c("ari", "hamming", "ami", "amri"))
}
\arguments{
\item{labels1}{label vector of clustering 1}

\item{labels2}{label vector of clustering 2}

\item{index.name}{string indicates the type of index to evaluate, one of c('ari','hamming',
'ami','amri')}
}
\value{
a scalar, calculated index value
}
\description{
Calculate the index between two clusterings
}
