% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{get_index_per_layer}
\alias{get_index_per_layer}
\title{Calculate the similarity between each layer of initial cluster tree and find MRTree results}
\usage{
get_index_per_layer(labelmat1, labelmat2, index.name = "ari")
}
\arguments{
\item{labelmat1}{n-by-m matrix, label matrix for cluster tree 1}

\item{labelmat2}{n-by-m matrix, label matrix for cluster tree 2}

\item{index.name}{string indicates the type of index to evaluate, one of c('ari','hamming',
'ami','amri')}
}
\value{
a numeric vector of length m, index per layer
}
\description{
Calculate the similarity between each layer of initial cluster tree and find MRTree results
}
