% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{get_set_D_size}
\alias{get_set_D_size}
\title{get the size of set D, {pairs that are in different clusters in low resolution clustering but in the
   same cluster in high resolution clustering}}
\usage{
get_set_D_size(labels.low.resolution, labels.high.resolution)
}
\arguments{
\item{labels.low.resolution}{vector of labels at lower resolution}

\item{labels.high.resolution}{vector of labels at higher resolution}
}
\value{
a scalar of size of set D
}
\description{
get the size of set D, {pairs that are in different clusters in low resolution clustering but in the
   same cluster in high resolution clustering}
}
