% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{get_similarity_from_labelmat}
\alias{get_similarity_from_labelmat}
\title{Get sample pairwise similarity matrix from the label matrix}
\usage{
get_similarity_from_labelmat(labelmat)
}
\arguments{
\item{labelmat}{a n-by-m labelmatrix, n samples, and m clusterings one per column}
}
\value{
a n-by-n similarity matrix
}
\description{
Get sample pairwise similarity matrix from the label matrix
}
