% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{get_similarity_from_tree}
\alias{get_similarity_from_tree}
\title{Calculate the pairwise similarity matrix between samples, from sample labels
and a phylo tree of labels.}
\usage{
get_similarity_from_tree(tree, labels)
}
\arguments{
\item{tree}{a phylo object giving the tree structure of labels}

\item{labels}{a label vector for n samples}
}
\value{
a n-by-n similarity matrix showing the similarity between samples
given the labels and the tree structure of labels
}
\description{
Calculate the pairwise similarity matrix between samples, from sample labels
and a phylo tree of labels.
}
