% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{getmode}
\alias{getmode}
\title{Get the mode in the vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{a vector of numeric or character}
}
\value{
a scalar or character representing the mode of the vector
}
\description{
Get the mode in the vector
}
