% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{grid_interpolant_next}
\alias{grid_interpolant_next}
\title{Grid intepolation using the next value}
\usage{
grid_interpolant_next(grid, x, y)
}
\arguments{
\item{grid}{a vector of new xs for which the ys will be generated by intepolation}

\item{x}{a vector of reference x where y is given}

\item{y}{a vector of value corresponding to x}
}
\value{
a vector of intepolated y values for the grid of x
}
\description{
Grid intepolation using the next value
}
