% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{hamming.distance}
\alias{hamming.distance}
\title{Hamming Distances of Vectors}
\usage{
hamming.distance(x, y)
}
\arguments{
\item{x}{a vector or matrix.}

\item{y}{an optional vector.}
}
\value{
distance matrix
}
\description{
f both x and y are vectors, hamming.distance returns the Hamming distance (number of different elements) between this two vectors. If x is a matrix, the Hamming distances between the rows of x are computed and y is ignored.
}
