% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{knn_impute}
\alias{knn_impute}
\title{Impute the missing labels by K nearest neighbor}
\usage{
knn_impute(labelmat)
}
\arguments{
\item{labelmat}{n-by-m sample by clustering label matrix}

\item{k}{number of nearest neighbor to consider}
}
\value{
Imputed labelmatrix
}
\description{
Impute the missing labels by K nearest neighbor
}
