% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{label_onehot}
\alias{label_onehot}
\title{one-hot code the label vector}
\usage{
label_onehot(labels, K = NULL)
}
\arguments{
\item{labels}{a vector of labels}

\item{K}{number of different categories, by default is the number of unique labels}
}
\value{
a n-by-K one-hot encoded binary matrix
}
\description{
one-hot code the label vector
}
