% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{label_to_membership}
\alias{label_to_membership}
\title{Convert label vector to membership matrix}
\usage{
label_to_membership(labels, label.names = NULL)

label_to_membership(labels, label.names = NULL)
}
\arguments{
\item{labels}{a vector of labels from K distint classes}

\item{labels.names}{(optional) alternative label names, used for naming columns of membership matrix}
}
\value{
an n-by-K binary membership matrix

an n-by-K binary membership matrix
}
\description{
Convert label vector to membership matrix

Convert label vector to membership matrix
}
