% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{linear_sampling}
\alias{linear_sampling}
\title{Linear sampling.}
\usage{
linear_sampling(n.res, res.min = 0.001, res.max = 3)
}
\arguments{
\item{n.res}{integer, number of resolutions to sample}

\item{res.min}{scalar, lower bound}

\item{res.max}{scalar, upper bound}
}
\value{
a vector of sampled parameters
}
\description{
Sample resolutions parameters linearly withing a range
}
