% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_both}
\alias{log_to_both}
\title{Log the message to both file and console}
\usage{
log_to_both(fn, logger = "ROOT", newfile = F)
}
\arguments{
\item{fn}{String, file name}

\item{logger}{String, logger name}

\item{newfile}{Boolean, new logging file}
}
\description{
Log the message to both file and console
}
