% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_console}
\alias{log_to_console}
\title{Log the message to console}
\usage{
log_to_console(logger = "ROOT")
}
\arguments{
\item{logger}{String, logger name}
}
\description{
Log the message to console
}
