% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_file}
\alias{log_to_file}
\title{Log the message to file}
\usage{
log_to_file(fn, logger = "ROOT", newfile = False)
}
\arguments{
\item{fn}{String, file name}

\item{logger}{String, logger name}

\item{newfile}{Boolean, new logging file}
}
\description{
Log the message to file
}
