% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{modularity}
\alias{modularity}
\title{Modularity function given the adjacency matrix}
\usage{
modularity(A, gamma = 1)
}
\arguments{
\item{A}{adjacency matrix}

\item{gamma}{scalar, parameter for modularity calculation}
}
\value{
numeric, calculated modularity value
}
\description{
Modularity function given the adjacency matrix
}
