% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{modularity_event_sampling}
\alias{modularity_event_sampling}
\title{Event sample method for resolution parameters of modularity clustering.}
\usage{
modularity_event_sampling(
  A,
  n.res,
  gamma.min = NULL,
  gamma.max = NULL,
  subsample = 200
)
}
\arguments{
\item{A}{n-by-n sample adjacency matrix}

\item{n.res}{total number of resolution parameters to sample}

\item{gamma.min}{lower bound for the range of resolution parameters to sample from}

\item{gamma.max}{upper bound for the range of resolution parameters to sample from}

\item{subsample}{integer, number of sample used for deriving the resolutions via subsampling (default 200). This procedure is adopted to save computation especially when n is large.}
}
\value{
a vector of length \code{n.res}, sampled resolution parameters.
}
\description{
Sampling the resolution parameter via Event Sampling algorithm proposed in Jeub, L. G., et al. (2018). "Multiresolution consensus clustering in networks." Scientific reports 8(1): 1-16.
}
