% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot MRtree results as a dendrogram. If reference labels are provided, a pie chart is
shown at each tree node, detailing the label proprotions.}
\usage{
plot_tree(
  labelmat,
  ref.labels = NULL,
  show.ref.labels = T,
  label.order = NULL,
  node.size = 0.2,
  cols = NULL,
  plot.piechart = T,
  tip.labels = NULL,
  tip.label.dist = 2,
  show.branch.labels = F,
  flip.branch = NULL,
  legend.title = "",
  bottom.margin = 25
)
}
\arguments{
\item{labelmat}{a n by m label matrix, in ouput of \code{mrtree} function, \code{labelmat.mrtree}}

\item{ref.labels}{a factor or characteristic vector specifying the reference labels of n data points}

\item{show.ref.labels}{boolean, whether to show the labels of major type at tree nodes and tips}

\item{label.order}{a vector specifying the order of labels for default colors}

\item{node.size}{scalar, the size of the pie chart / node}

\item{cols}{a vector of colors, one per each label}

\item{plot.piechart}{boolean, whether to draw the pie chart in the tree plot}

\item{tip.labels}{a vector of strings specifying the labels of tree leafs}

\item{tip.label.dist}{distance of the tip labels to the tree tips}

\item{show.branch.labels}{boolean, whether to show the branch labels for convenience of flipping branches}

\item{legend.title}{string as legend title, default is an empty string}

\item{bottom.margin}{size of the bottom margin, need to be adjusted to show the full labels}

\item{fip.branch}{a list of vectors each of size 2, indicating the branch labels to flip. Each time two branches are flipped.}
}
\description{
Plot MRtree results as a dendrogram. If reference labels are provided, a pie chart is
shown at each tree node, detailing the label proprotions.
}
