% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_tree_with_piechart}
\alias{plot_tree_with_piechart}
\title{Plot MRtree results as a dendrogram. If reference labels are provided, a pie chart is
shown at each tree node, detailing the label proprotions.}
\usage{
plot_tree_with_piechart(
  labelmat,
  ref.labels = NULL,
  factor.level.names = NULL,
  draw.pie.chart = TRUE,
  pie.size.offset = 1,
  show.ref.labels = T,
  tip.label.dist = 2,
  node.label.dist = 2,
  bottom.margin = 25
)
}
\arguments{
\item{labelmat}{a n by m label matrix, in ouput of \code{mrtree} function, \code{labelmat.mrtree}}

\item{ref.labels}{a factor or characteristic vector specifying the reference labels of n data points}

\item{factor.level.names}{a vector specifying the order in color the reference labels are shown}

\item{draw.pie.chart}{boolean, whether to draw the pie chart in the tree plot}

\item{pie.size.offset}{scalar offset the size of the pie chart (larger the smaller the pie chart is)}

\item{show.ref.labels}{boolean, whether to show the labels of major type at tree nodes and tips}

\item{tip.label.dist}{distance of the tip labels to the tree tips}

\item{node.label.dist}{distance of the node labels to the nodes}

\item{bottom.margin}{size of the bottom margin, need to be adjusted to show the full labels}
}
\description{
Plot MRtree results as a dendrogram. If reference labels are provided, a pie chart is
shown at each tree node, detailing the label proprotions.
}
