% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP plot}
\usage{
plot_umap(
  X = NULL,
  labels = NULL,
  pca = 50,
  n_components = 2,
  n_neighbors = 30,
  min_dist = 0.1,
  point.size = 0.3,
  alpha = 1,
  title = NULL,
  legend.name = "labels",
  cols = NULL,
  emb = NULL,
  seed = 0
)
}
\arguments{
\item{X}{n-by-p expression matrix}

\item{labels}{vector of sample labels}

\item{pca}{umap parameter, dimension of pca}

\item{n_components}{umap parameter, dimension of low dimensional embedding space (default 2)}

\item{n_neighbors}{umap parameter, number of neighbors for nearest neighbor graph}

\item{min_dist}{umap parameter, minimum distance of point to its nearest neighbor in the
embedding space.}

\item{point.size}{numeric scalar, point size in the plot}

\item{alpha}{numeric, transparency of the points in the plot}

\item{title}{string, title of the plot}

\item{legend.name}{string, legend name}

\item{cols}{vector of colors, length should the same as cardinality of labels}

\item{emb}{embedding of the UMAP if provided}

\item{seed}{random seed}
}
\value{
A list of \describe{
\item{p} umap plot
\item{emb} umap embedding matrix
}
}
\description{
UMAP plot
}
