% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotContTable}
\alias{plotContTable}
\title{Plot the confusion matrix via heatmap, with true labels in columns}
\usage{
plotContTable(
  est_label,
  true_label,
  true_label_order = NULL,
  est_label_order = NULL,
  short.names = NULL,
  xlab = "Reference",
  ylab = NULL
)
}
\arguments{
\item{est_label}{vector of estimated labels (rows)}

\item{true_label}{vector of actual labels (columns)}

\item{true_label_order}{order of column labels if provided}

\item{est_label_order}{order of row labels if provided}

\item{short.names}{short names of the true labels to be shown}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
ggplot object
}
\description{
Plot the confusion matrix via heatmap, with true labels in columns
}
