% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree_functions.R
\name{prune_paths}
\alias{prune_paths}
\title{Remove the paths that include the same node.end but node.start different}
\usage{
prune_paths(paths, node.start, node.end)
}
\arguments{
\item{paths}{Initial paths to be pruned}

\item{node.start}{the label of the starting node}

\item{node.end}{the label of the ending node}
}
\value{
A matrix whose rows contains the remaining paths after prunning
}
\description{
Remove the paths that include the same node.end but node.start different
}
