% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{rep_mat}
\alias{rep_mat}
\title{repeat the vector as a matrix}
\usage{
rep_mat(x, k)
}
\arguments{
\item{x}{input vector to repeat}

\item{k}{number of time to repeat}
}
\value{
a matrix of k columns
}
\description{
repeat the vector as a matrix
}
