% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.HAC}
\alias{sc_clustering.HAC}
\title{Perform Single Cell data clustering using Hierarchical Agglomerative clustering (HAC)}
\usage{
sc_clustering.HAC(
  exprs,
  n.pcs = 10,
  labels = NULL,
  method = "complete",
  type = c("count", "log"),
  scale.factor = 10000,
  verbose = F
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{n.pcs}{integer, number of principal components to use for calculating similarity and performn HAC}

\item{labels}{vector of string/factor, if provided, HAC is performed starting from clusters}

\item{method}{string, similarity method, one of 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA),
'mcquitty' (= WPGMA), 'median' (= WPGMC) or 'centroid' (= UPGMC).}

\item{type}{string, type of the expression matrix, choices are 'count' and 'log', and default by 'counts}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{verbose}{boolean, whether to print messages}
}
\value{
an hclust object containing the clustering results
}
\description{
Perform Single Cell data clustering using Hierarchical Agglomerative clustering (HAC)
}
