% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.sc3}
\alias{sc_clustering.sc3}
\title{Perform Single Cell data clustering using SC3 clustering pipeline}
\usage{
sc_clustering.sc3(
  exprs,
  Ks,
  type = c("counts", "normcounts"),
  colData = NULL,
  rowData = NULL,
  estimate.k = F,
  scale.factor = 10000,
  build.hierarchical.tree = F,
  ...
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{Ks}{vector of resolution, number of clusters}

\item{type}{string, type of the expression matrix, choices are 'counts', 'normcounts' and 'logcounts', and default by 'counts'}

\item{colData}{a dataframe containing cell informations}

\item{rowData}{a dataframe containing gene informations}

\item{estimate.k}{boolean, whether to estimate optimal number of clusters by sc3}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{build.hierarchical.tree}{boolean, whether to obtain HAC results using SC3}

\item{...}{other parameters input to \code{sc3} function}
}
\value{
a list containing \describe{
\item{sce}{an SingleCellExperiment object containing all the clustering results}
\item{hc.tree}{hclust object resulted from hierarchical agglomerative clustering
using obtained by SC3}
}
}
\description{
Perform Single Cell data clustering using SC3 clustering pipeline
}
