% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.seurat}
\alias{sc_clustering.seurat}
\title{Perform Single Cell data clustering using Seurat}
\usage{
sc_clustering.seurat(
  counts,
  resolutions,
  metadata = NULL,
  min.cells = 0,
  min.features = 0,
  scale.factor = 10000,
  vars.to.regress = NULL,
  find.variable.features = T,
  npcs = 40,
  seurat.graph.algorithm = 1,
  build.hierarchical.tree = FALSE,
  return.seurat.object = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{counts}{n.genes-by-n.cells count matrix}

\item{resolutions}{vector of clustering resolution paramers (input for FindClusters)}

\item{metadata}{a data frame containing all the cell informations (equivalent to colData)}

\item{min.cells}{integer, include features detected in at least this number cells}

\item{min.features}{integer, include cells where at least this number features are detected}

\item{scale.factor}{scalar, sets the scale factor for cell-level normalization}

\item{vars.to.regress}{a vector of strings, variables to regress out}

\item{find.variable.features}{T: find highly variable features using 'mean.var.plot',
if >1: find variable features using 'vst' to get find.variable.features genes. F: not
genes selection is performed}

\item{npcs}{integer, number of principal component to calculate. The PCS are used for build
neighbor graph}

\item{seurat.graph.algorithm}{algorithm for modularity optimization (1 = original
Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm;
4 = Leiden algorithm). Leiden requires the leidenalg python.}

\item{build.hierarchical.tree}{boolean, whether to build hierarchical tree using HAC
from Seurat clusters with \code{max(resolution)}.}

\item{return.seurat.object}{boolean, whether to return Seurat object. Save memory by setting it to \code{FALSE}.}

\item{verbose}{boolean, whether to print messages}
}
\value{
a list containing \describe{
    \item{seurat.clusters}{a data frame containing all clusteringS as columns with
    prefix 'RNA_snn_res.'}
    \item{obj}{Seurat object before clustering is performed}
    \item{hc.tree}{hclust object resulted from hierarchical agglomerative clustering
    using Seurat clusters from \code{max(resolutions)}}
}
}
\description{
Perform Single Cell data clustering using Seurat
}
