% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.simlr}
\alias{sc_clustering.simlr}
\title{Perform Single Cell data clustering using SIMLR}
\usage{
sc_clustering.simlr(
  exprs,
  Ks,
  type = c("count", "log"),
  estimate.k = F,
  scale.factor = 10^4,
  cores.ratio = 0.9,
  column.prefix = "simlr_"
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{Ks}{vector of resolution, number of clusters}

\item{type}{string, type of the expression matrix, choices are 'count' and 'log', and default by 'counts'}

\item{estimate.k}{whether to estimate optimal number of clusters by SIMLR}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{cores.ratio}{ratio of total number of cores used for calculation}
}
\value{
an SingleCellExperiment object containing all the clustering results
}
\description{
Perform Single Cell data clustering using SIMLR
}
