% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.soup}
\alias{sc_clustering.soup}
\title{Perform Single Cell data clustering using SOUP}
\usage{
sc_clustering.soup(
  exprs,
  Ks,
  type = c("count", "log"),
  estimate.k = FALSE,
  pure.prop = 0.5,
  ext.prop = NULL,
  scale.factor = 10000,
  column.prefix = "soup_"
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{Ks}{vector of resolution, number of clusters}

\item{type}{string, type of the expression matrix, choices are 'count' and 'log', and default by 'counts'}

\item{estimate.k}{boolean, whether to estimate optimal number of clusters by soup}

\item{pure.prop}{the proportion of pure cells, SOUP parameter}

\item{ext.prop}{the proportion of extreme neighbors for each cell, SOUP parameter}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{column.prefix}{string, output column prefix, default 'soup_'}
}
\value{
a list containing \describe{
    \item{labelmat}{a data frame, columns are clusterings for each resolution specified}
    \item{est.k}{integer, estimated number of clusters}
}
}
\description{
Perform Single Cell data clustering using SOUP
}
