% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.tsne_kmeans}
\alias{sc_clustering.tsne_kmeans}
\title{Perform Single Cell data clustering using TSNE+kmeans}
\usage{
sc_clustering.tsne_kmeans(
  exprs,
  Ks,
  type = c("count", "log"),
  estimate.k = FALSE,
  scale.factor = 10000,
  dimensions = 2,
  perplexity = 20,
  column.prefix = "tsnekmeans_"
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{Ks}{vector of resolution, number of clusters}

\item{type}{string, type of the expression matrix, choices are 'count' and 'log', and default by 'counts'}

\item{estimate.k}{whether to estimate optimal number of clusters by ADPclust}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{dimensions}{integer, 2 or 3, TSNE output dimensionality}

\item{perplexity}{integer, Perplexity parameter for TSNE}

\item{column.prefix}{string of output column prefix, default 'tsnekmeans_'}
}
\value{
a list containing \describe{
\item{labelmat}{a data frame, columns are clusterings for each resolution specified}
\item{est.k}{integer, estimated number of clusters}
}
}
\description{
Perform Single Cell data clustering using TSNE+kmeans
}
