% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{sc_clustering.umap_kmeans}
\alias{sc_clustering.umap_kmeans}
\title{Perform Single Cell data clustering using UMAP+kmeans}
\usage{
sc_clustering.umap_kmeans(
  exprs,
  Ks,
  type = c("count", "log"),
  estimate.k = FALSE,
  subsample = F,
  subsample.ratio = 0.9,
  scale.factor = 10000,
  n.neighbors = 30,
  n.components = 2,
  column.prefix = "umapkmeans_",
  n.cores = parallel::detectCores() - 4
)
}
\arguments{
\item{exprs}{n.genes-by-n.cells expression matrix}

\item{Ks}{vector of resolution, number of clusters}

\item{type}{string, type of the expression matrix, choices are 'count' and 'log', and default by 'counts'}

\item{estimate.k}{boolean whether to estimate optimal number of clusters by ADPclust}

\item{subsample}{whether perform subsampling for each clustering}

\item{subsample.ratio}{ratio of subsampled size to the total sample size (applicable when subsample=TRUE)}

\item{scale.factor}{scalar sets the scale factor for cell-level normalization}

\item{n.neighbors}{integer, number of neighbors to consider when build neighbor graph, default 30}

\item{n.components}{integer, UMAP output dimensionality, default 2}

\item{column.prefix}{string, output column prefix, default 'tsnekmeans_'}

\item{n.cores}{number of cores used for parallel computation}
}
\value{
a list containing \describe{
\item{labelmat}{a data frame, columns are clusterings for each resolution specified}
\item{est.k}{integer, estimated number of clusters}
}
}
\description{
Perform Single Cell data clustering using UMAP+kmeans
}
