% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{set_log_level}
\alias{set_log_level}
\title{Set the log threshold}
\usage{
set_log_level(level = "INFO", logger = "ROOT")
}
\arguments{
\item{level}{log streshold}

\item{logger}{logger name}
}
\description{
Set the log threshold
}
