% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{stability_plot}
\alias{stability_plot}
\title{Stability analysis that measures the similarity between the initial flat clusterins and reconciled tree
at each resolution. Similarity can be ARI or AMRI}
\usage{
stability_plot(out, index.name = "ari")
}
\arguments{
\item{out}{MRtree output, list containing the labelmat.mrtree and labelmat.flat}

\item{index.name}{string, measurement of similarity, that can be 'ari' or 'amri'}
}
\value{
a dataframe showing the similarity per resolution
}
\description{
A line plot is generated showing the similarity across resolutions, measuring the clustering stability
for each resolution
}
