% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{table_to_edgelist}
\alias{table_to_edgelist}
\title{Convert the start end table to an edgelist}
\usage{
table_to_edgelist(labels.start, labels.end)
}
\arguments{
\item{labels.start}{label of the start nodes}

\item{labels.end}{labels of the end nodes}
}
\value{
and edge list \describe{
\item{start}{label of start node}
\item{end}{label of end node}
\item{count}{number of data points in the edge}
}
}
\description{
Convert the start end table to an edgelist
}
