% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_simulation_functions.R
\name{tree1}
\alias{tree1}
\title{Generate tree structure}
\usage{
tree1(plot = FALSE)
}
\arguments{
\item{plot}{boolean, whether to generate plot}
}
\value{
A list containing \describe{
    \item{phylo.tree}{Generated tree saved in a phylo object}
    \item{fig}{figure that shows the tree structure}
}
}
\description{
Generate tree structure
}
