% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{tree_to_labelmat.hclust}
\alias{tree_to_labelmat.hclust}
\title{Convert hclust tree to label matrix}
\usage{
\method{tree_to_labelmat}{hclust}(tree, Ks = NULL)
}
\arguments{
\item{tree}{hclust object representing a dendrogram}

\item{Ks}{number of clusters per layer}
}
\value{
a label matrix
}
\description{
Convert hclust tree to label matrix
}
