% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{tree_to_labelmat.phylo}
\alias{tree_to_labelmat.phylo}
\title{Convert phylo tree to labelmatrix}
\usage{
\method{tree_to_labelmat}{phylo}(tree, Ks = NULL)
}
\arguments{
\item{tree}{phylo tree}

\item{Ks}{number of clusters per layer}
}
\value{
a label matrix
}
\description{
Convert phylo tree to labelmatrix
}
