% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{tree_to_labelmat}
\alias{tree_to_labelmat}
\title{Convert phylo_tree to label matrix}
\usage{
tree_to_labelmat(tree, ...)
}
\arguments{
\item{tree}{phylo object or hclust object}

\item{...}{other parameters}
}
\description{
Convert phylo_tree to label matrix
}
