% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{unique_ind}
\alias{unique_ind}
\title{Get the unique values in the matrix}
\usage{
unique_ind(A)
}
\arguments{
\item{matrix}{A}
}
\value{
a list of \describe{
\item{C}{sorted unique values (increaing order)}
\item{ia}{a vector of index such that \code{C = c(A)[ia]}}
\item{ic}{index vector such that \code{c(A) = C[ic]}}
}
}
\description{
Get the unique values in the matrix
}
