% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{augment_path}
\alias{augment_path}
\title{Augment the path to include on additional node as alternative to each layer}
\usage{
augment_path(paths, layers, aug.name = -1)
}
\description{
Augment the path to include on additional node as alternative to each layer
}
