% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree_functions.R
\name{cost}
\alias{cost}
\title{Calculate the cost generate by adding the edge = (start, end) and removing all the other conflicting edge}
\usage{
cost(node.start, node.end, paths, labelmat, labelmat.in.paths, weights)
}
\arguments{
\item{node.start}{start node of the edge (out-vertex)}

\item{node.end}{end node of the edge (in-vertex)}

\item{paths}{viable path}

\item{labelmat.in.paths}{assigned labels of data points to the optimimum paths}

\item{weights}{weights used in loss function (all ones by default)}
}
\value{
a scalar representing the cost
}
\description{
Calculate the cost generate by adding the edge = (start, end) and removing all the other conflicting edge
}
