% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering_methods.R
\name{create_sce_from_normcounts}
\alias{create_sce_from_normcounts}
\title{Create SingleCellExperiment object from normalized data}
\usage{
create_sce_from_normcounts(normcounts, col.data, row.data = NULL)
}
\arguments{
\item{normcounts}{gene-by-cell matrix of normalized expressions}

\item{col.data}{a dataframe containing cell informations}

\item{row.data}{a dataframe containing gene informations}
}
\value{
a SingleCellExperiment object
}
\description{
Create SingleCellExperiment object from normalized data
}
