% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate the defult colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{integer, number of colors}
}
\value{
n colors
}
\description{
Generate the defult colors
}
