% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtree.R
\name{mrtree}
\alias{mrtree}
\alias{mrtree.matrix}
\alias{mrtree.data.frame}
\alias{mrtree.SingleCellExperiment}
\alias{mrtree.Seurat}
\title{MRtree main function}
\usage{
mrtree(x, ...)

\method{mrtree}{matrix}(
  labelmat,
  prefix = NULL,
  suffix = NULL,
  max.k = Inf,
  consensus = F,
  sample.weighted = F,
  verbose = F,
  n.cores = parallel::detectCores() - 1
)

\method{mrtree}{data.frame}(x, prefix, suffix = NULL, ...)

\method{mrtree}{SingleCellExperiment}(x, prefix = "sc3_", suffix = "_clusters", ...)

\method{mrtree}{Seurat}(x, prefix = "RNA_snn_res.", suffix = NULL, ...)
}
\arguments{
\item{x}{object containing multi-resolution clustering results
\strong{data source}
building a reconciled hierarchical clustering tree requires information about which cluster each sample has been assinged
to at different resolutions. This information can be supplied in various forms, as a matrix, data.frame or more specialised
object. In all cases the object provided must contain numeric columns with the naming structure \code{PXS} where \code{P} is
a prefix indicating that the column contains clustering information, \code{X} is a numeric value indicating the clustering
resolution and \code{S} is any additional suffix to be removed. For \code{SingleCellExperiment} objects this information must
be in the \code{colData} slot and for \code{Seurat} objects it must be in the \code{meta.data} slot. For all objects except matrices
any additional columns can be used as aesthetics, for matrices an additional metadata data.frame can be supplied if
required.}

\item{prefix}{srting indicating columns containing clustering information}

\item{suffix}{string at the end of column names containing clustering information}

\item{max.k}{the maximum resolution (number of clusters) to consider in building the tree}

\item{consensus}{boolean, whether to perform consensus clustering within the clusterings with the same number of clusters.}

\item{verbose}{boolean, whether to show logmsgs}

\item{n.cores}{integer, number of cores for parallel computation}
}
\value{
A list containing  \describe{
\item{labelmat.mrtree}{The Reconciled tree saved as a label matrix, with duplicated layers omited.}
\item{labelmat.recon}{The full reconciled tree save in a label matrix}
\item{labelmat.flat}{The initial flat clustering (cluster tree) as input for MRtree algorithm}
\item{resolutions}{The corresponding clustering resolution of the initial cluster tree}
\item{paths}{The unique path in the resulted reconciled hierarchical cluster tree}
}
}
\description{
The multi-resolution clusters is input into \code{mrtree}, and the output is the constructed hierarchical cluster tree.
}
