% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_simulation_functions.R
\name{random_tree}
\alias{random_tree}
\title{Generate a random tree with even number of tips}
\usage{
random_tree(n)
}
\arguments{
\item{n}{number of tips}
}
\description{
Generate a random tree with even number of tips
}
